add `withLibraryFields` as an export option
if set to true, datasets and dataset references will
contain library fields if applicable. In the service
API, we sometimes want to know if a referenced dataset
in another dataset is from a library (e.g. a process
in a product system), we can this then configure
directly in the export.