bug-fix: recursive types are not correctly handled by the batch import
the batch import creates entities until a batch is full. then, it inserts
the entities and sets them to `visited` after this. when an entity points
to another entity in the batch, it will look into the import cache for
this. The import cache does not find it, and it will use the normal JSON
import outside the batch for creating it. After this, we can have two
instances (or more) of the same thing in the database, depending on the
import order etc.