analysis groups: filter out double links
when there are double links (same process, provider, and flow) in the
system, these are resolved to a single matrix entry, adding up the
respective amounts. When constructing the sub-graph for the analysis
groups, we need to filter out these double links then.

this filtering makes the sub-graph construction a bit slower, but we
removed the recursice calls when adding a path to the sub-graph
instead.